embedder/pyproject.toml
[project]
name = "whatsapp-embedder"
version = "0.1.0"
description = "Embedding worker for WhatsApp Bridge - stores message vectors in pgvector"
requires-python = ">=3.11"
dependencies = [
    "psycopg2-binary>=2.9",
    "sentence-transformers>=3.0",
]

[project.scripts]
whatsapp-embedder = "whatsapp_embedder.worker:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/whatsapp_embedder"]
